[update] SSR optimisation rendering
