+++ next.config.ts
@@ -10,6 +10,9 @@ const nextConfig: NextConfig = {
       },
     ],
   },
+  compiler: {
+    styledComponents: true,
+  },
 };
 
 export default nextConfig;
